Change interaction hierarchy again
