+++ core/src/main/java/discord4j/core/event/dispatch/DispatchHandlers.java
@@ -1,4 +1,21 @@
 /*
+ *  This file is part of Discord4J.
+ *
+ * Discord4J is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Discord4J is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Discord4J. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/*
  * This file is part of Discord4J.
  *
  * Discord4J is free software: you can redistribute it and/or modify
@@ -279,6 +296,8 @@ public class DispatchHandlers implements DispatchEventMapper {
                         return Mono.just(new ComponentInteractionEvent(gateway, context.getShardInfo(), interaction));
                 }
 
+            case APPLICATION_COMMAND_AUTOCOMPLETE:
+                return Mono.just(new ChatInputAutoCompleteEvent(gateway, context.getShardInfo(), interaction));
             default:
                 return Mono.just(new InteractionCreateEvent(gateway, context.getShardInfo(), interaction));
         }
