+++ core/src/main/java/discord4j/core/spec/InteractionApplicationCommandCallbackSpecGenerator.java
@@ -18,7 +18,7 @@
 package discord4j.core.spec;
 
 import discord4j.core.event.domain.interaction.ComponentInteractionEvent;
-import discord4j.core.event.domain.interaction.InteractionCreateEvent;
+import discord4j.core.event.domain.interaction.DeferrableInteractionEvent;
 import discord4j.core.object.component.LayoutComponent;
 import discord4j.core.object.entity.Message;
 import discord4j.discordjson.json.InteractionApplicationCommandCallbackData;
@@ -70,7 +70,7 @@ interface InteractionApplicationCommandCallbackSpecGenerator extends Spec<Intera
 abstract class InteractionApplicationCommandCallbackReplyMonoGenerator extends Mono<Void>
         implements InteractionApplicationCommandCallbackSpecGenerator {
 
-    abstract InteractionCreateEvent event();
+    abstract DeferrableInteractionEvent event();
 
     @Override
     public void subscribe(CoreSubscriber<? super Void> actual) {
