+++ core/src/test/java/discord4j/core/ExampleInteractions.java
@@ -1,4 +1,21 @@
 /*
+ *  This file is part of Discord4J.
+ *
+ * Discord4J is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Discord4J is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Discord4J. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/*
  * This file is part of Discord4J.
  *
  * Discord4J is free software: you can redistribute it and/or modify
@@ -39,7 +56,7 @@ import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 
 /**
- * Example to showcase how {@link InteractionCreateEvent#deferReply()}, {@link InteractionCreateEvent#reply()},
+ * Example to showcase how {@link DeferrableInteractionEvent#deferReply()}, {@link DeferrableInteractionEvent#reply()},
  * {@link ComponentInteractionEvent#deferEdit()} and {@link ComponentInteractionEvent#edit()} work for chat input, user,
  * message commands and component interactions.
  */
