+++ rest/src/main/java/discord4j/rest/util/InteractionResponseType.java
@@ -1,4 +1,21 @@
 /*
+ *  This file is part of Discord4J.
+ *
+ * Discord4J is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Discord4J is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Discord4J. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+/*
  * This file is part of Discord4J.
  *
  * Discord4J is free software: you can redistribute it and/or modify
@@ -43,7 +60,12 @@ public enum InteractionResponseType {
     /**
      * For components, edit the message the component was attached to
      */
-    UPDATE_MESSAGE(7);
+    UPDATE_MESSAGE(7),
+
+    /**
+     * Respond to an autocomplete interaction with suggested choices
+     */
+    APPLICATION_COMMAND_AUTOCOMPLETE_RESULT(8);
 
     /**
      * The underlying value as represented by Discord.
